Split Template API functions to logical groups; Update docs
